Example of data

Set of data useful for trying the package.

exampleData

Format

An object of class list of length 3.

Source

These data are simulated and was generated using the serious game "PlantBreedGame" available on GitHub: https://github.com/timflutre/PlantBreedGame

Flutre, T., Diot, J., and David, J. (2019). PlantBreedGame: A Serious Game that Puts Students in the Breeder’s Seat. Crop Science. DOI 10.2135/cropsci2019.03.0183le

Details

This object is a named list containing 3 elements:

genotypes (data.frame): Genotypic data encoded in allele doses of 100 fictitious individuals with 3333 SNP markers. These individuals have 10 chromosomes of length 10^6 bases pairs.

snpCoord (data.frame): Coordinates of the 3333 individuals' markers. This data.frame contains 3 columns: chr, pos and SNPid.

snpEffects (numeric): "true" effects of the 3333 individuals' markers about a fictitious quantitative trait based on an additive architecture.

Examples

exampleData$genotypes

#>      snp00006 snp00009 snp00011 snp00018 snp00026 snp00031 snp00035 snp00036
#>      snp00049 snp00052 snp00075 snp00076 snp00087 snp00101 snp00102 snp00103
#>      snp00111 snp00113 snp00116 snp00117 snp00130 snp00131 snp00141 snp00144
#>      snp00145 snp00166 snp00172 snp00173 snp00177 snp00185 snp00190 snp00191
#>      snp00201 snp00202 snp00205 snp00227 snp00239 snp00249 snp00255 snp00256
#>      snp00259 snp00324 snp00349 snp00439 snp00472 snp00546 snp00552 snp00562
#>      snp00563 snp00565 snp00583 snp00585 snp00589 snp00594 snp00602 snp00603
#>      snp00621 snp00629 snp00634 snp00640 snp00642 snp00643 snp00655 snp00662
#>      snp00669 snp00670 snp00689 snp00694 snp00702 snp00711 snp00712 snp00715
#>      snp00717 snp00721 snp00722 snp00724 snp00742 snp00758 snp00766 snp00796
#>      snp00801 snp00808 snp00810 snp00812 snp00816 snp00826 snp00830 snp00832
#>      snp00834 snp00836 snp00842 snp00847 snp00855 snp00868 snp00875 snp00877
#>      snp00888 snp00889 snp00891 snp00892 snp00894 snp00913 snp00943 snp00948
#>      snp00953 snp00955 snp00957 snp00958 snp00967 snp00979 snp00980 snp00982
#>      snp00992 snp00996 snp01003 snp01041 snp01047 snp01052 snp01062 snp01078
#>      snp01097 snp01101 snp01118 snp01120 snp01161 snp01227 snp01229 snp01245
#>      snp01246 snp01265 snp01267 snp01271 snp01282 snp01283 snp01294 snp01319
#>      snp01324 snp01325 snp01328 snp01330 snp01353 snp01354 snp01362 snp01368
#>      snp01373 snp01374 snp01382 snp01384 snp01392 snp01399 snp01404 snp01423
#>      snp01424 snp01435 snp01436 snp01480 snp01489 snp01499 snp01500 snp01504
#>      snp01512 snp01514 snp01516 snp01517 snp01531 snp01535 snp01536 snp01540
#>      snp01549 snp01563 snp01599 snp01620 snp01648 snp01651 snp01657 snp01668
#>      snp01683 snp01686 snp01687 snp01688 snp01697 snp01712 snp01714 snp01733
#>      snp01779 snp01790 snp01798 snp01808 snp01831 snp01838 snp01842 snp01843
#>      snp01855 snp01859 snp01871 snp01877 snp01878 snp01882 snp01885 snp01894
#>      snp01905 snp01907 snp01917 snp01930 snp01935 snp01936 snp01937 snp01954
#>      snp01957 snp01958 snp01987 snp01992 snp02017 snp02021 snp02024 snp02029
#>      snp02036 snp02044 snp02049 snp02067 snp02072 snp02074 snp02121 snp02126
#>      snp02130 snp02139 snp02140 snp02146 snp02149 snp02150 snp02152 snp02159
#>      snp02160 snp02167 snp02178 snp02185 snp02190 snp02196 snp02206 snp02214
#>      snp02217 snp02225 snp02228 snp02231 snp02232 snp02234 snp02240 snp02249
#>      snp02261 snp02262 snp02303 snp02328 snp02347 snp02352 snp02375 snp02384
#>      snp02399 snp02402 snp02406 snp02411 snp02426 snp02463 snp02469 snp02482
#>      snp02503 snp02506 snp02507 snp02512 snp02520 snp02523 snp02535 snp02545
#>      snp02553 snp02560 snp02569 snp02572 snp02586 snp02590 snp02604 snp02610
#>      snp02618 snp02666 snp02674 snp02684 snp02693 snp02706 snp02712 snp02715
#>      snp02730 snp02757 snp02759 snp02800 snp02831 snp02850 snp02852 snp02853
#>      snp02859 snp02874 snp02890 snp02894 snp02905 snp02916 snp02917 snp02933
#>      snp02938 snp02940 snp02945 snp02946 snp02954 snp02961 snp02962 snp02996
#>      snp03006 snp03019 snp03021 snp03023 snp03052 snp03071 snp03081 snp03094
#>      snp03098 snp03103 snp03105 snp03133 snp03143 snp03153 snp03165 snp03167
#>      snp03209 snp03211 snp03221 snp03225 snp03247 snp03248 snp03254 snp03259
#>      snp03262 snp03268 snp03270 snp03284 snp03323 snp03380 snp03384 snp03386
#>      snp03408 snp03420 snp03475 snp03476 snp03488 snp03494 snp03503 snp03517
#>      snp03519 snp03533 snp03542 snp03551 snp03559 snp03568 snp03573 snp03579
#>      snp03580 snp03597 snp03612 snp03622 snp03626 snp03642 snp03647 snp03657
#>      snp03668 snp03669 snp03679 snp03683 snp03684 snp03700 snp03704 snp03711
#>      snp03728 snp03732 snp03735 snp03747 snp03751 snp03755 snp03758 snp03760
#>      snp03763 snp03767 snp03779 snp03795 snp03820 snp03821 snp03833 snp03848
#>      snp03859 snp03881 snp03891 snp03895 snp03896 snp03904 snp03910 snp03923
#>      snp03931 snp03943 snp03952 snp03956 snp03973 snp03974 snp03976 snp03977
#>      snp03982 snp03994 snp04000 snp04004 snp04007 snp04013 snp04019 snp04053
#>      snp04056 snp04066 snp04070 snp04079 snp04118 snp04133 snp04140 snp04141
#>      snp04142 snp04152 snp04157 snp04160 snp04164 snp04218 snp04231 snp04254
#>      snp04262 snp04269 snp04281 snp04293 snp04346 snp04358 snp04359 snp04365
#>      snp04367 snp04377 snp04394 snp04395 snp04404 snp04408 snp04424 snp04425
#>      snp04430 snp04431 snp04445 snp04480 snp04486 snp04487 snp04505 snp04537
#>      snp04538 snp04542 snp04549 snp04570 snp04593 snp04669 snp04684 snp04692
#>      snp04699 snp04701 snp04707 snp04716 snp04718 snp04723 snp04762 snp04767
#>      snp04778 snp04807 snp04816 snp04818 snp04823 snp04825 snp04900 snp04903
#>      snp04924 snp04929 snp04939 snp04973 snp04989 snp05003 snp05004 snp05026
#>      snp05031 snp05037 snp05040 snp05052 snp05060 snp05066 snp05068 snp05071
#>      snp05083 snp05087 snp05115 snp05120 snp05121 snp05135 snp05137 snp05140
#>      snp05213 snp05214 snp05230 snp05258 snp05277 snp05278 snp05287 snp05291
#>      snp05293 snp05296 snp05303 snp05318 snp05335 snp05340 snp05344 snp05368
#>      snp05383 snp05392 snp05393 snp05396 snp05420 snp05439 snp05443 snp05448
#>      snp05464 snp05470 snp05477 snp05490 snp05505 snp05508 snp05509 snp05514
#>      snp05540 snp05558 snp05562 snp05568 snp05571 snp05606 snp05656 snp05718
#>      snp05764 snp05773 snp05809 snp05817 snp05825 snp05854 snp05871 snp05873
#>      snp05893 snp05896 snp05908 snp05938 snp05940 snp05964 snp05970 snp05976
#>      snp05977 snp05978 snp05981 snp05992 snp06006 snp06014 snp06020 snp06082
#>      snp06087 snp06109 snp06132 snp06163 snp06176 snp06196 snp06197 snp06205
#>      snp06206 snp06224 snp06234 snp06239 snp06294 snp06305 snp06307 snp06320
#>      snp06334 snp06335 snp06386 snp06423 snp06435 snp06438 snp06439 snp06493
#>      snp06499 snp06510 snp06514 snp06515 snp06534 snp06542 snp06550 snp06563
#>      snp06568 snp06574 snp06576 snp06579 snp06619 snp06630 snp06637 snp06647
#>      snp06660 snp06672 snp06677 snp06678 snp06681 snp06690 snp06705 snp06707
#>      snp06710 snp06719 snp06726 snp06736 snp06737 snp06745 snp06747 snp06755
#>      snp06771 snp06795 snp06810 snp06812 snp06919 snp06923 snp06924 snp06934
#>      snp06950 snp06962 snp06967 snp06968 snp06975 snp06981 snp06987 snp06989
#>      snp07015 snp07019 snp07022 snp07032 snp07034 snp07036 snp07048 snp07054
#>      snp07056 snp07057 snp07077 snp07081 snp07086 snp07091 snp07095 snp07096
#>      snp07108 snp07113 snp07118 snp07124 snp07139 snp07141 snp07151 snp07154
#>      snp07163 snp07168 snp07172 snp07174 snp07184 snp07197 snp07209 snp07215
#>      snp07219 snp07240 snp07243 snp07254 snp07257 snp07263 snp07271 snp07272
#>      snp07274 snp07280 snp07285 snp07287 snp07290 snp07304 snp07312 snp07336
#>      snp07349 snp07356 snp07358 snp07360 snp07367 snp07369 snp07372 snp07374
#>      snp07379 snp07389 snp07412 snp07445 snp07463 snp07464 snp07546 snp07553
#>      snp07573 snp07579 snp07601 snp07608 snp07612 snp07625 snp07629 snp07644
#>      snp07650 snp07653 snp07654 snp07674 snp07678 snp07683 snp07687 snp07702
#>      snp07708 snp07709 snp07720 snp07723 snp07733 snp07752 snp07757 snp07759
#>      snp07762 snp07763 snp07770 snp07794 snp07797 snp07803 snp07813 snp07825
#>      snp07828 snp07834 snp07846 snp07848 snp07852 snp07862 snp07865 snp07870
#>      snp07878 snp07895 snp07903 snp07904 snp07905 snp07908 snp07918 snp07921
#>      snp07933 snp07938 snp07954 snp07955 snp07956 snp07963 snp07975 snp07977
#>      snp07979 snp07987 snp07988 snp08001 snp08002 snp08015 snp08016 snp08017
#>      snp08020 snp08035 snp08116 snp08123 snp08125 snp08126 snp08139 snp08179
#>      snp08180 snp08224 snp08233 snp08240 snp08247 snp08268 snp08286 snp08298
#>      snp08307 snp08318 snp08319 snp08324 snp08330 snp08331 snp08336 snp08343
#>      snp08352 snp08380 snp08386 snp08387 snp08391 snp08397 snp08402 snp08410
#>      snp08421 snp08438 snp08453 snp08462 snp08465 snp08473 snp08479 snp08485
#>      snp08488 snp08507 snp08510 snp08513 snp08516 snp08532 snp08538 snp08544
#>      snp08546 snp08554 snp08561 snp08570 snp08582 snp08586 snp08592 snp08595
#>      snp08600 snp08612 snp08622 snp08634 snp08635 snp08636 snp08642 snp08646
#>      snp08661 snp08667 snp08673 snp08677 snp08730 snp08746 snp08753 snp08764
#>      snp08778 snp08780 snp08787 snp08798 snp08801 snp08813 snp08829 snp08835
#>      snp08837 snp08840 snp08846 snp08850 snp08852 snp08870 snp08873 snp08878
#>      snp08880 snp08913 snp08917 snp08920 snp08927 snp08938 snp08946 snp08950
#>      snp08952 snp08956 snp08959 snp08968 snp08980 snp09062 snp09069 snp09078
#>      snp09079 snp09081 snp09084 snp09085 snp09094 snp09098 snp09099 snp09106
#>      snp09137 snp09152 snp09160 snp09161 snp09177 snp09197 snp09204 snp09206
#>      snp09208 snp09210 snp09216 snp09218 snp09223 snp09239 snp09248 snp09254
#>      snp09259 snp09260 snp09261 snp09272 snp09286 snp09295 snp09301 snp09311
#>      snp09317 snp09330 snp09334 snp09349 snp09353 snp09354 snp09364 snp09413
#>      snp09417 snp09419 snp09421 snp09425 snp09431 snp09447 snp09465 snp09478
#>      snp09486 snp09489 snp09494 snp09496 snp09505 snp09518 snp09536 snp09540
#>      snp09542 snp09553 snp09598 snp09605 snp09606 snp09630 snp09635 snp09641
#>      snp09642 snp09658 snp09676 snp09691 snp09697 snp09698 snp09702 snp09704
#>      snp09711 snp09716 snp09717 snp09730 snp09731 snp09758 snp09767 snp09768
#>      snp09786 snp09789 snp09790 snp09795 snp09800 snp09816 snp09817 snp09831
#>      snp09842 snp09845 snp09859 snp09861 snp09862 snp09863 snp09871 snp09876
#>      snp09910 snp09914 snp09916 snp09920 snp09930 snp09943 snp09945 snp09948
#>      snp09954 snp09957 snp09965 snp09968 snp09970 snp09986 snp09994 snp10007
#>      snp10008 snp10023 snp10024 snp10026 snp10043 snp10111 snp10125 snp10129
#>      snp10131 snp10132 snp10160 snp10179 snp10196 snp10221 snp10226 snp10260
#>      snp10271 snp10278 snp10280 snp10282 snp10302 snp10321 snp10322 snp10332
#>      snp10337 snp10344 snp10346 snp10347 snp10354 snp10361 snp10368 snp10370
#>      snp10383 snp10385 snp10392 snp10393 snp10395 snp10404 snp10411 snp10414
#>      snp10415 snp10420 snp10422 snp10428 snp10434 snp10441 snp10452 snp10465
#>      snp10472 snp10475 snp10486 snp10497 snp10503 snp10510 snp10512 snp10519
#>      snp10533 snp10540 snp10545 snp10549 snp10557 snp10567 snp10573 snp10594
#>      snp10595 snp10597 snp10615 snp10620 snp10638 snp10646 snp10653 snp10656
#>      snp10687 snp10707 snp10714 snp10716 snp10719 snp10722 snp10726 snp10728
#>      snp10732 snp10735 snp10738 snp10743 snp10749 snp10751 snp10752 snp10755
#>      snp10764 snp10775 snp10778 snp10786 snp10790 snp10805 snp10832 snp10859
#>      snp10877 snp10890 snp10891 snp10892 snp10923 snp10924 snp10925 snp10930
#>      snp10940 snp10953 snp10958 snp10963 snp10965 snp10966 snp10979 snp10988
#>      snp10996 snp10997 snp11006 snp11019 snp11033 snp11043 snp11045 snp11050
#>      snp11070 snp11090 snp11091 snp11099 snp11133 snp11168 snp11190 snp11193
#>      snp11197 snp11201 snp11209 snp11213 snp11217 snp11224 snp11231 snp11233
#>      snp11248 snp11265 snp11266 snp11286 snp11306 snp11315 snp11316 snp11320
#>      snp11332 snp11336 snp11341 snp11346 snp11353 snp11368 snp11370 snp11375
#>      snp11376 snp11386 snp11390 snp11402 snp11405 snp11409 snp11410 snp11416
#>      snp11417 snp11418 snp11425 snp11448 snp11465 snp11490 snp11492 snp11520
#>      snp11533 snp11550 snp11560 snp11580 snp11585 snp11596 snp11597 snp11600
#>      snp11612 snp11615 snp11640 snp11641 snp11646 snp11672 snp11678 snp11693
#>      snp11696 snp11698 snp11699 snp11725 snp11729 snp11766 snp11773 snp11777
#>      snp11791 snp11824 snp11830 snp11837 snp11842 snp11856 snp11874 snp11879
#>      snp11901 snp11903 snp11908 snp11910 snp11914 snp11919 snp11921 snp11924
#>      snp11945 snp11946 snp11953 snp11959 snp11962 snp11964 snp11968 snp11971
#>      snp11977 snp11985 snp11989 snp12005 snp12053 snp12062 snp12063 snp12089
#>      snp12161 snp12171 snp12173 snp12174 snp12184 snp12197 snp12199 snp12211
#>      snp12222 snp12228 snp12256 snp12268 snp12270 snp12295 snp12316 snp12318
#>      snp12330 snp12335 snp12343 snp12378 snp12384 snp12388 snp12389 snp12390
#>      snp12393 snp12396 snp12400 snp12403 snp12447 snp12449 snp12455 snp12462
#>      snp12463 snp12464 snp12495 snp12497 snp12513 snp12524 snp12533 snp12538
#>      snp12539 snp12546 snp12592 snp12597 snp12602 snp12610 snp12616 snp12633
#>      snp12645 snp12654 snp12661 snp12670 snp12673 snp12680 snp12688 snp12702
#>      snp12705 snp12707 snp12713 snp12717 snp12719 snp12721 snp12731 snp12732
#>      snp12733 snp12740 snp12762 snp12772 snp12783 snp12787 snp12795 snp12803
#>      snp12804 snp12814 snp12819 snp12821 snp12826 snp12831 snp12840 snp12848
#>      snp12851 snp12859 snp12881 snp12885 snp12888 snp12926 snp12929 snp12933
#>      snp12938 snp12943 snp12947 snp12948 snp12950 snp12961 snp12968 snp12993
#>      snp13011 snp13066 snp13078 snp13085 snp13099 snp13100 snp13111 snp13114
#>      snp13127 snp13174 snp13178 snp13186 snp13224 snp13228 snp13237 snp13245
#>      snp13255 snp13257 snp13258 snp13259 snp13269 snp13277 snp13291 snp13305
#>      snp13307 snp13336 snp13349 snp13350 snp13358 snp13371 snp13380 snp13405
#>      snp13408 snp13410 snp13412 snp13531 snp13545 snp13555 snp13574 snp13586
#>      snp13587 snp13588 snp13597 snp13631 snp13634 snp13641 snp13651 snp13659
#>      snp13664 snp13665 snp13672 snp13675 snp13690 snp13692 snp13698 snp13702
#>      snp13703 snp13721 snp13736 snp13745 snp13755 snp13778 snp13797 snp13842
#>      snp13845 snp13847 snp13851 snp13852 snp13855 snp13885 snp13886 snp13893
#>      snp13895 snp13904 snp13905 snp13912 snp13914 snp13916 snp13940 snp13961
#>      snp13979 snp13992 snp14006 snp14008 snp14009 snp14010 snp14014 snp14020
#>      snp14025 snp14028 snp14035 snp14038 snp14040 snp14083 snp14095 snp14098
#>      snp14100 snp14117 snp14120 snp14121 snp14122 snp14145 snp14150 snp14183
#>      snp14192 snp14262 snp14263 snp14270 snp14278 snp14290 snp14304 snp14316
#>      snp14317 snp14365 snp14374 snp14381 snp14384 snp14385 snp14392 snp14399
#>      snp14417 snp14427 snp14436 snp14450 snp14462 snp14471 snp14476 snp14479
#>      snp14482 snp14493 snp14497 snp14505 snp14513 snp14515 snp14524 snp14543
#>      snp14549 snp14562 snp14576 snp14579 snp14580 snp14595 snp14597 snp14697
#>      snp14706 snp14711 snp14729 snp14748 snp14757 snp14777 snp14783 snp14788
#>      snp14800 snp14803 snp14852 snp14881 snp14883 snp14889 snp14890 snp14891
#>      snp14892 snp14894 snp14911 snp14919 snp14927 snp14944 snp14952 snp14959
#>      snp14962 snp14963 snp14970 snp14981 snp15008 snp15032 snp15052 snp15057
#>      snp15085 snp15089 snp15141 snp15155 snp15157 snp15159 snp15244 snp15257
#>      snp15259 snp15264 snp15266 snp15273 snp15283 snp15294 snp15309 snp15357
#>      snp15362 snp15365 snp15372 snp15374 snp15378 snp15396 snp15400 snp15409
#>      snp15412 snp15418 snp15424 snp15435 snp15440 snp15445 snp15448 snp15460
#>      snp15470 snp15481 snp15485 snp15488 snp15492 snp15498 snp15501 snp15504
#>      snp15506 snp15508 snp15513 snp15525 snp15559 snp15583 snp15591 snp15592
#>      snp15596 snp15598 snp15599 snp15606 snp15621 snp15639 snp15720 snp15728
#>      snp15731 snp15742 snp15743 snp15745 snp15747 snp15749 snp15901 snp15921
#>      snp15938 snp15951 snp15953 snp15959 snp15962 snp15963 snp15965 snp15986
#>      snp15998 snp16001 snp16008 snp16010 snp16012 snp16014 snp16022 snp16024
#>      snp16029 snp16036 snp16037 snp16040 snp16051 snp16063 snp16064 snp16090
#>      snp16099 snp16105 snp16125 snp16129 snp16130 snp16143 snp16149 snp16151
#>      snp16152 snp16153 snp16176 snp16181 snp16194 snp16205 snp16218 snp16221
#>      snp16241 snp16245 snp16247 snp16248 snp16267 snp16269 snp16270 snp16283
#>      snp16298 snp16301 snp16306 snp16307 snp16314 snp16322 snp16339 snp16342
#>      snp16361 snp16362 snp16384 snp16395 snp16404 snp16405 snp16407 snp16448
#>      snp16449 snp16452 snp16460 snp16461 snp16462 snp16468 snp16470 snp16474
#>      snp16479 snp16509 snp16514 snp16528 snp16601 snp16602 snp16609 snp16613
#>      snp16620 snp16621 snp16624 snp16628 snp16630 snp16633 snp16637 snp16647
#>      snp16699 snp16712 snp16721 snp16738 snp16739 snp16741 snp16744 snp16777
#>      snp16779 snp16800 snp16803 snp16807 snp16808 snp16810 snp16812 snp16828
#>      snp16877 snp16896 snp16910 snp16914 snp16927 snp16934 snp16938 snp16943
#>      snp16945 snp16948 snp16952 snp16983 snp16992 snp16994 snp16996 snp17005
#>      snp17029 snp17035 snp17044 snp17061 snp17097 snp17103 snp17118 snp17125
#>      snp17126 snp17152 snp17153 snp17160 snp17161 snp17166 snp17168 snp17172
#>      snp17176 snp17192 snp17195 snp17214 snp17218 snp17229 snp17235 snp17272
#>      snp17278 snp17289 snp17290 snp17314 snp17318 snp17320 snp17337 snp17344
#>      snp17348 snp17356 snp17357 snp17362 snp17368 snp17375 snp17378 snp17379
#>      snp17385 snp17386 snp17395 snp17405 snp17408 snp17409 snp17428 snp17440
#>      snp17458 snp17462 snp17464 snp17467 snp17486 snp17487 snp17488 snp17492
#>      snp17496 snp17498 snp17499 snp17501 snp17502 snp17515 snp17529 snp17539
#>      snp17565 snp17566 snp17572 snp17587 snp17592 snp17594 snp17596 snp17613
#>      snp17630 snp17631 snp17638 snp17647 snp17654 snp17656 snp17665 snp17667
#>      snp17686 snp17690 snp17691 snp17697 snp17704 snp17706 snp17718 snp17721
#>      snp17725 snp17745 snp17760 snp17768 snp17770 snp17812 snp17846 snp17867
#>      snp17868 snp17870 snp17891 snp17896 snp17903 snp17904 snp17911 snp17915
#>      snp17916 snp17920 snp17921 snp17938 snp17943 snp17954 snp17967 snp17980
#>      snp17989 snp17990 snp17994 snp18015 snp18034 snp18046 snp18051 snp18054
#>      snp18071 snp18081 snp18096 snp18097 snp18106 snp18107 snp18135 snp18138
#>      snp18141 snp18154 snp18178 snp18190 snp18192 snp18194 snp18201 snp18205
#>      snp18208 snp18209 snp18218 snp18220 snp18222 snp18227 snp18229 snp18232
#>      snp18235 snp18241 snp18259 snp18271 snp18275 snp18285 snp18286 snp18293
#>      snp18295 snp18302 snp18303 snp18323 snp18338 snp18344 snp18353 snp18355
#>      snp18365 snp18382 snp18385 snp18387 snp18388 snp18389 snp18392 snp18397
#>      snp18403 snp18407 snp18416 snp18425 snp18432 snp18436 snp18445 snp18457
#>      snp18462 snp18465 snp18497 snp18504 snp18535 snp18560 snp18562 snp18594
#>      snp18603 snp18609 snp18613 snp18627 snp18666 snp18671 snp18681 snp18682
#>      snp18684 snp18695 snp18769 snp18775 snp18779 snp18783 snp18808 snp18885
#>      snp18888 snp18944 snp19014 snp19038 snp19070 snp19071 snp19073 snp19076
#>      snp19081 snp19088 snp19110 snp19115 snp19119 snp19121 snp19137 snp19145
#>      snp19148 snp19150 snp19161 snp19186 snp19190 snp19195 snp19205 snp19235
#>      snp19241 snp19242 snp19243 snp19245 snp19250 snp19260 snp19264 snp19267
#>      snp19279 snp19280 snp19285 snp19298 snp19307 snp19332 snp19336 snp19341
#>      snp19350 snp19353 snp19355 snp19357 snp19395 snp19405 snp19406 snp19413
#>      snp19417 snp19436 snp19445 snp19449 snp19452 snp19457 snp19463 snp19475
#>      snp19479 snp19480 snp19499 snp19513 snp19519 snp19524 snp19527 snp19528
#>      snp19540 snp19554 snp19558 snp19561 snp19565 snp19581 snp19588 snp19592
#>      snp19597 snp19603 snp19617 snp19647 snp19654 snp19679 snp19681 snp19683
#>      snp19692 snp19704 snp19726 snp19780 snp19811 snp19837 snp19854 snp19863
#>      snp19867 snp19883 snp19893 snp19905 snp19908 snp19922 snp19960 snp19976
#>      snp19979 snp19986 snp19991 snp19998 snp20008 snp20012 snp20039 snp20045
#>      snp20050 snp20052 snp20078 snp20091 snp20092 snp20099 snp20111 snp20177
#>      snp20178 snp20179 snp20186 snp20221 snp20304 snp20314 snp20326 snp20340
#>      snp20347 snp20353 snp20405 snp20409 snp20422 snp20434 snp20448 snp20453
#>      snp20454 snp20466 snp20499 snp20500 snp20503 snp20530 snp20537 snp20541
#>      snp20544 snp20550 snp20558 snp20581 snp20586 snp20588 snp20589 snp20592
#>      snp20598 snp20613 snp20621 snp20630 snp20639 snp20654 snp20659 snp20662
#>      snp20667 snp20669 snp20680 snp20683 snp20687 snp20706 snp20710 snp20711
#>      snp20722 snp20750 snp20770 snp20773 snp20776 snp20785 snp20799 snp20808
#>      snp20812 snp20826 snp20834 snp20840 snp20847 snp20851 snp20854 snp20856
#>      snp20858 snp20866 snp20870 snp20873 snp20877 snp20882 snp20886 snp20896
#>      snp20908 snp20940 snp20952 snp21005 snp21151 snp21199 snp21222 snp21321
#>      snp21331 snp21335 snp21336 snp21366 snp21374 snp21384 snp21386 snp21410
#>      snp21417 snp21419 snp21421 snp21435 snp21452 snp21479 snp21484 snp21488
#>      snp21490 snp21514 snp21607 snp21757 snp21775 snp21778 snp21780 snp21783
#>      snp21785 snp21789 snp21790 snp21793 snp21805 snp21807 snp21808 snp21905
#>      snp21916 snp21922 snp21925 snp21932 snp21937 snp21951 snp21954 snp22016
#>      snp22048 snp22049 snp22050 snp22073 snp22101 snp22105 snp22114 snp22131
#>      snp22148 snp22156 snp22190 snp22191 snp22203 snp22215 snp22226 snp22239
#>      snp22240 snp22244 snp22259 snp22261 snp22271 snp22282 snp22299 snp22313
#>      snp22316 snp22329 snp22353 snp22360 snp22372 snp22376 snp22380 snp22390
#>      snp22407 snp22427 snp22428 snp22434 snp22445 snp22457 snp22458 snp22465
#>      snp22491 snp22496 snp22501 snp22507 snp22514 snp22516 snp22519 snp22523
#>      snp22541 snp22572 snp22578 snp22581 snp22589 snp22601 snp22613 snp22621
#>      snp22626 snp22628 snp22684 snp22699 snp22702 snp22706 snp22730 snp22743
#>      snp22772 snp22773 snp22785 snp22804 snp22835 snp22848 snp22854 snp22874
#>      snp22881 snp22893 snp22925 snp22947 snp22949 snp22973 snp22979 snp22981
#>      snp22984 snp22987 snp22990 snp22993 snp23027 snp23059 snp23070 snp23092
#>      snp23172 snp23177 snp23190 snp23201 snp23204 snp23214 snp23264 snp23301
#>      snp23355 snp23383 snp23399 snp23403 snp23416 snp23417 snp23429 snp23437
#>      snp23456 snp23480 snp23488 snp23490 snp23551 snp23556 snp23560 snp23562
#>      snp23568 snp23570 snp23572 snp23578 snp23597 snp23646 snp23653 snp23654
#>      snp23685 snp23690 snp23697 snp23699 snp23711 snp23734 snp23756 snp23765
#>      snp23777 snp23782 snp23790 snp23806 snp23825 snp23826 snp23838 snp23846
#>      snp23847 snp23849 snp23851 snp23861 snp23871 snp23875 snp23885 snp23890
#>      snp23896 snp23902 snp23911 snp23923 snp23927 snp23932 snp23934 snp23938
#>      snp23951 snp23973 snp23975 snp23990 snp24002 snp24078 snp24096 snp24116
#>      snp24138 snp24162 snp24165 snp24181 snp24188 snp24199 snp24201 snp24225
#>      snp24231 snp24263 snp24287 snp24289 snp24296 snp24300 snp24305 snp24335
#>      snp24343 snp24345 snp24349 snp24427 snp24439 snp24445 snp24456 snp24459
#>      snp24461 snp24464 snp24478 snp24481 snp24482 snp24490 snp24492 snp24499
#>      snp24510 snp24546 snp24557 snp24560 snp24564 snp24568 snp24576 snp24580
#>      snp24597 snp24605 snp24607 snp24631 snp24641 snp24643 snp24652 snp24664
#>      snp24710 snp24715 snp24717 snp24718 snp24720 snp24779 snp24789 snp24800
#>      snp24801 snp24804 snp24807 snp24809 snp24824 snp24838 snp24857 snp24912
#>      snp24913 snp24914 snp24921 snp24932 snp24936 snp24947 snp24956 snp24966
#>      snp24978 snp24982 snp24983 snp25011 snp25019 snp25020 snp25038 snp25049
#>      snp25054 snp25055 snp25058 snp25077 snp25092 snp25096 snp25118 snp25264
#>      snp25281 snp25284 snp25286 snp25289 snp25291 snp25304 snp25310 snp25313
#>      snp25316 snp25317 snp25318 snp25319 snp25320 snp25323 snp25327 snp25351
#>      snp25361 snp25367 snp25370 snp25373 snp25377 snp25380 snp25384 snp25396
#>      snp25404 snp25407 snp25415 snp25423 snp25435 snp25443 snp25447 snp25448
#>      snp25459 snp25470 snp25473 snp25488 snp25491 snp25498 snp25502 snp25511
#>      snp25518 snp25529 snp25531 snp25546 snp25552 snp25567 snp25585 snp25586
#>      snp25590 snp25607 snp25622 snp25629 snp25632 snp25635 snp25641 snp25665
#>      snp25689 snp25698 snp25699 snp25701 snp25739 snp25745 snp25759 snp25769
#>      snp25805 snp25806 snp25809 snp25810 snp25814 snp25846 snp25849 snp25852
#>      snp25855 snp25862 snp25864 snp25873 snp25879 snp25880 snp25891 snp25899
#>      snp25900 snp25911 snp25918 snp25930 snp25933 snp25934 snp25935 snp25958
#>      snp25959 snp25974 snp25979 snp25996 snp26004 snp26019 snp26029 snp26030
#>      snp26058 snp26061 snp26065 snp26072 snp26073 snp26082 snp26088 snp26091
#>      snp26092 snp26094 snp26096 snp26128 snp26218 snp26224 snp26231 snp26241
#>      snp26245 snp26250 snp26263 snp26264 snp26267 snp26269 snp26271 snp26275
#>      snp26279 snp26282 snp26294 snp26295 snp26311 snp26317 snp26327 snp26444
#>      snp26478 snp26489 snp26500 snp26512 snp26516 snp26517 snp26523 snp26542
#>      snp26578 snp26584 snp26594 snp26606 snp26622 snp26633 snp26647 snp26649
#>      snp26656 snp26694 snp26699 snp26701 snp26712 snp26723 snp26725 snp26728
#>      snp26729 snp26737 snp26739 snp26744 snp26745 snp26746 snp26754 snp26755
#>      snp26763 snp26801 snp26813 snp26831 snp26838 snp26840 snp26861 snp26867
#>      snp26880 snp26885 snp26887 snp26898 snp26937 snp26967 snp26989 snp26993
#>      snp26999 snp27013 snp27015 snp27016 snp27040 snp27051 snp27061 snp27069
#>      snp27081 snp27092 snp27095 snp27098 snp27099 snp27101 snp27102 snp27105
#>      snp27106 snp27136 snp27150 snp27165 snp27170 snp27176 snp27178 snp27201
#>      snp27204 snp27216 snp27219 snp27237 snp27239 snp27240 snp27244 snp27245
#>      snp27249 snp27251 snp27254 snp27257 snp27261 snp27275 snp27279 snp27287
#>      snp27291 snp27298 snp27302 snp27304 snp27306 snp27315 snp27322 snp27325
#>      snp27409 snp27412 snp27420 snp27423 snp27426 snp27428 snp27447 snp27450
#>      snp27456 snp27459 snp27462 snp27467 snp27469 snp27472 snp27475 snp27484
#>      snp27491 snp27493 snp27500 snp27510 snp27516 snp27529 snp27531 snp27571
#>      snp27577 snp27588 snp27592 snp27609 snp27620 snp27635 snp27672 snp27683
#>      snp27685 snp27687 snp27691 snp27697 snp27704 snp27709 snp27745 snp27797
#>      snp27806 snp27819 snp27822 snp27826 snp27833 snp27836 snp27837 snp27843
#>      snp27845 snp27848 snp27849 snp27851 snp27859 snp27861 snp27873 snp27875
#>      snp27879 snp27883 snp27887 snp27891 snp27892 snp27893 snp27906 snp27908
#>      snp27910 snp27912 snp27937 snp27940 snp27951 snp27953 snp27955 snp27964
#>      snp27974 snp27980 snp27989 snp27993 snp28002 snp28003 snp28007 snp28021
#>      snp28031 snp28032 snp28033 snp28055 snp28095 snp28096 snp28100 snp28107
#>      snp28126 snp28157 snp28180 snp28183 snp28185 snp28191 snp28198 snp28208
#>      snp28212 snp28214 snp28220 snp28237 snp28238 snp28242 snp28247 snp28253
#>      snp28257 snp28271 snp28273 snp28281 snp28286 snp28309 snp28323 snp28327
#>      snp28337 snp28352 snp28354 snp28355 snp28356 snp28369 snp28382 snp28385
#>      snp28401 snp28409 snp28418 snp28430 snp28439 snp28445 snp28451 snp28454
#>      snp28465 snp28495 snp28498 snp28528 snp28538 snp28559 snp28572 snp28595
#>      snp28602 snp28616 snp28619 snp28633 snp28638 snp28642 snp28644 snp28661
#>      snp28680 snp28694 snp28724 snp28763 snp28765 snp28769 snp28776 snp28781
#>      snp28787 snp28794 snp28799 snp28800 snp28801 snp28806 snp28807 snp28822
#>      snp28826 snp28842 snp28852 snp28865 snp28869 snp28870 snp28874 snp28877
#>      snp28894 snp28896 snp28902 snp28911 snp28912 snp28916 snp28917 snp28919
#>      snp28931 snp28936 snp28975 snp28992 snp28994 snp28999 snp29018 snp29019
#>      snp29024 snp29036 snp29039 snp29047 snp29055 snp29062 snp29076 snp29082
#>      snp29085 snp29088 snp29092 snp29098 snp29106 snp29108 snp29109 snp29110
#>      snp29126 snp29128 snp29138 snp29143 snp29145 snp29150 snp29156 snp29159
#>      snp29162 snp29169 snp29174 snp29175 snp29178 snp29189 snp29193 snp29200
#>      snp29202 snp29207 snp29224 snp29233 snp29234 snp29243 snp29247 snp29254
#>      snp29255 snp29316 snp29327 snp29343 snp29346 snp29350 snp29357 snp29360
#>      snp29361 snp29367 snp29369 snp29433 snp29447 snp29451 snp29477 snp29479
#>      snp29495 snp29497 snp29503 snp29523 snp29541 snp29542 snp29580 snp29588
#>      snp29590 snp29595 snp29628 snp29632 snp29657 snp29669 snp29728 snp29737
#>      snp29745 snp29750 snp29752 snp29770 snp29801 snp29808 snp29809 snp29812
#>      snp29821 snp29831 snp29832 snp29835 snp29843 snp29845 snp29850 snp29858
#>      snp29873 snp29881 snp29884 snp29888 snp29889 snp29891 snp29897 snp29902
#>      snp29903 snp29916 snp29917 snp29928 snp29940 snp29957 snp29975 snp29977
#>      snp29978 snp30009 snp30012 snp30020 snp30022 snp30027 snp30048 snp30060
#>      snp30068 snp30077 snp30078 snp30085 snp30089 snp30103 snp30135 snp30136
#>      snp30137 snp30143 snp30148 snp30150 snp30154 snp30167 snp30169 snp30190
#>      snp30199 snp30205 snp30207 snp30224 snp30228 snp30245 snp30266 snp30277
#>      snp30288 snp30300 snp30304 snp30329 snp30345 snp30346 snp30358 snp30385
#>      snp30391 snp30397 snp30407 snp30425 snp30454 snp30465 snp30542 snp30563
#>      snp30566 snp30581 snp30620 snp30621 snp30635 snp30639 snp30641 snp30674
#>      snp30677 snp30678 snp30680 snp30694 snp30708 snp30710 snp30716 snp30723
#>      snp30725 snp30733 snp30743 snp30754 snp30757 snp30775 snp30838 snp30859
#>      snp30864 snp30871 snp30872 snp30894 snp30964 snp30976 snp31000 snp31011
#>      snp31012 snp31015 snp31035 snp31070 snp31110 snp31114 snp31120 snp31142
#>      snp31155 snp31164 snp31182 snp31189 snp31191 snp31194 snp31213 snp31215
#>      snp31221 snp31223 snp31227 snp31234 snp31239 snp31250 snp31251 snp31271
#>      snp31274 snp31276 snp31285 snp31286 snp31304 snp31305 snp31311 snp31330
#>      snp31356 snp31363 snp31375 snp31385 snp31387 snp31394 snp31398 snp31399
#>      snp31400 snp31409 snp31414 snp31425 snp31435 snp31441 snp31448 snp31452
#>      snp31463 snp31472 snp31473 snp31475 snp31479 snp31487 snp31488 snp31489
#>      snp31490 snp31494 snp31527 snp31532 snp31571 snp31574 snp31578 snp31584
#>      snp31606 snp31616 snp31620 snp31622 snp31626 snp31634 snp31640 snp31641
#>      snp31643 snp31652 snp31654 snp31660 snp31667 snp31689 snp31703 snp31714
#>      snp31792 snp31820 snp31877 snp31880 snp31888 snp31893 snp31915 snp31919
#>      snp31920 snp31924 snp31944 snp31953 snp31955 snp31956 snp31961 snp31977
#>      snp31979 snp31991 snp31996 snp32028 snp32031 snp32034 snp32035 snp32043
#>      snp32054 snp32057 snp32059 snp32060 snp32064 snp32067 snp32069 snp32071
#>      snp32083 snp32084 snp32096 snp32104 snp32110 snp32116 snp32126 snp32140
#>      snp32145 snp32156 snp32158 snp32162 snp32168 snp32189 snp32195 snp32198
#>      snp32204 snp32213 snp32218 snp32223 snp32226 snp32230 snp32236 snp32258
#>      snp32270 snp32309 snp32317 snp32318 snp32335 snp32341 snp32350 snp32362
#>      snp32376 snp32382 snp32388 snp32396 snp32407 snp32411 snp32413 snp32415
#>      snp32420 snp32433 snp32449 snp32460 snp32495 snp32516 snp32518 snp32521
#>      snp32532 snp32533 snp32539 snp32545 snp32549 snp32559 snp32595 snp32617
#>      snp32639 snp32641 snp32642 snp32665 snp32672 snp32678 snp32681 snp32696
#>      snp32698 snp32730 snp32749 snp32772 snp32780 snp32788 snp32800 snp32847
#>      snp32874 snp32892 snp32914 snp32921 snp32930 snp32955 snp32957 snp32961
#>      snp32963 snp32965 snp32973 snp32974 snp32989 snp33003 snp33032 snp33035
#>      snp33037 snp33041 snp33059 snp33060 snp33070 snp33077 snp33083 snp33093
#>      snp33145 snp33159 snp33238 snp33265 snp33268 snp33269 snp33272 snp33275
#>      snp33277 snp33293 snp33300 snp33313 snp33314 snp33321 snp33344 snp33349
#>      snp33368 snp33372 snp33392 snp33406 snp33408 snp33417 snp33508 snp33510
#>      snp33511 snp33514 snp33516 snp33520 snp33544 snp33554 snp33558 snp33559
#>      snp33563 snp33569 snp33579 snp33591 snp33593 snp33597 snp33607 snp33620
#>      snp33621 snp33626 snp33647 snp33650 snp33651 snp33681 snp33683 snp33687
#>      snp33697 snp33720 snp33731 snp33735 snp33739 snp33741 snp33747 snp33756
#>      snp33757 snp33764 snp33765 snp33779 snp33843 snp33851 snp33862 snp33879
#>      snp33892 snp33906 snp33914 snp33919 snp33925 snp33930 snp33961 snp33962
#>      snp33965 snp33985 snp33987 snp33992 snp33995 snp34000 snp34006 snp34008
#>      snp34019 snp34026 snp34043 snp34052 snp34064 snp34066 snp34077 snp34078
#>      snp34080 snp34081 snp34199 snp34201 snp34296 snp34301 snp34321 snp34329
#>      snp34330 snp34332 snp34337 snp34340 snp34341 snp34344 snp34351 snp34362
#>      snp34366 snp34383 snp34386 snp34397 snp34405 snp34424 snp34431 snp34444
#>      snp34463 snp34494 snp34495 snp34497 snp34504 snp34531 snp34534 snp34538
#>      snp34541 snp34554 snp34562 snp34598 snp34612 snp34614 snp34617 snp34633
#>      snp34638 snp34651 snp34656 snp34678 snp34689 snp34704 snp34718 snp34749
#>      snp34752 snp34763 snp34765 snp34770 snp34793 snp34803 snp34810 snp34813
#>      snp34817 snp34837 snp34842 snp34849 snp34857 snp34869 snp34875 snp34883
#>      snp34891 snp34896 snp34914 snp34920 snp34921 snp34922 snp34929 snp34931
#>      snp34932 snp34936 snp34939 snp34947 snp34962 snp34983 snp34989 snp34994
#>      snp35016 snp35024 snp35048 snp35050 snp35059 snp35069 snp35070 snp35082
#>      snp35086 snp35103 snp35106 snp35117 snp35120 snp35123 snp35134 snp35151
#>      snp35186 snp35188 snp35202 snp35203 snp35207
#>  [ reached 'max' / getOption("max.print") -- omitted 100 rows ]

exampleData$snpCoord

#>       chr    pos    SNPid
#> 1   Chr10 223757 snp32341
#> 2   Chr01 937638 snp03760
#> 3   Chr05 933373 snp17846
#> 4   Chr02 664717 snp06006
#> 5   Chr04 216465 snp11597
#> 6   Chr01 654763 snp02674
#> 7   Chr03 232247 snp07908
#> 8   Chr04 867517 snp13851
#> 9   Chr05 369642 snp15460
#> 10  Chr01 181658 snp00721
#> 11  Chr06  67587 snp18397
#> 12  Chr09 235658 snp28992
#> 13  Chr03 353491 snp08324
#> 14  Chr10 479029 snp33238
#> 15  Chr08 200198 snp25327
#> 16  Chr08 609223 snp26622
#> 17  Chr02 603635 snp05773
#> 18  Chr01 230126 snp00948
#> 19  Chr09 252739 snp29092
#> 20  Chr08 682433 snp26861
#> 21  Chr05 879017 snp17572
#> 22  Chr06 444649 snp19245
#> 23  Chr08 475317 snp26218
#> 24  Chr06 444496 snp19241
#> 25  Chr09 153969 snp28595
#> 26  Chr01 420637 snp01620
#> 27  Chr08 239778 snp25511
#> 28  Chr08 927577 snp27691
#> 29  Chr01 467620 snp01790
#> 30  Chr05  60296 snp14462
#> 31  Chr03 480593 snp08778
#> 32  Chr05 912499 snp17745
#> 33  Chr03  27670 snp07263
#> 34  Chr06  14703 snp18154
#> 35  Chr05 840248 snp17464
#> 36  Chr09 772286 snp30639
#> 37  Chr10 195761 snp32236
#> 38  Chr04 816276 snp13641
#> 39  Chr04 549535 snp12762
#> 40  Chr07 917099 snp24287
#> 41  Chr03 482792 snp08801
#> 42  Chr05 781216 snp17195
#> 43  Chr07 673039 snp23556
#> 44  Chr08 203157 snp25351
#> 45  Chr01 605192 snp02469
#> 46  Chr07 607531 snp23301
#> 47  Chr09 777115 snp30674
#> 48  Chr03 788907 snp09994
#> 49  Chr07 205038 snp22261
#> 50  Chr02 363128 snp05060
#> 51  Chr10 878723 snp34689
#> 52  Chr09 221971 snp28911
#> 53  Chr10 844037 snp34541
#> 54  Chr08 694034 snp26887
#> 55  Chr01 221759 snp00913
#> 56  Chr06 884502 snp20873
#> 57  Chr06 620729 snp19863
#> 58  Chr07 194597 snp22239
#> 59  Chr08 314756 snp25809
#> 60  Chr03 728796 snp09758
#> 61  Chr07  30999 snp21514
#> 62  Chr07  24269 snp21452
#> 63  Chr04 991282 snp14183
#> 64  Chr08 748542 snp27098
#> 65  Chr08 958600 snp27879
#> 66  Chr07 299436 snp22519
#> 67  Chr09  35061 snp28126
#> 68  Chr06 545417 snp19681
#> 69  Chr01 888624 snp03597
#> 70  Chr08 199261 snp25318
#> 71  Chr06 904371 snp20952
#> 72  Chr07 211354 snp22282
#> 73  Chr05 772351 snp17160
#> 74  Chr06  51096 snp18285
#> 75  Chr08 950982 snp27806
#> 76  Chr04 391474 snp12228
#> 77  Chr04 974144 snp14122
#> 78  Chr10 142316 snp31996
#> 79  Chr03 190065 snp07733
#> 80  Chr07  27291 snp21488
#> 81  Chr09 257365 snp29126
#> 82  Chr03 731452 snp09767
#> 83  Chr07  90689 snp21780
#> 84  Chr02 660335 snp05992
#> 85  Chr05 823126 snp17357
#> 86  Chr02 234454 snp04699
#> 87  Chr01 346006 snp01353
#> 88  Chr04 660451 snp13099
#> 89  Chr04 884455 snp13914
#> 90  Chr06 842739 snp20669
#> 91  Chr05 789303 snp17218
#> 92  Chr04 456266 snp12403
#> 93  Chr10 335724 snp32749
#> 94  Chr06 805177 snp20448
#> 95  Chr03 870478 snp10322
#> 96  Chr08 838102 snp27531
#> 97  Chr10 584955 snp33544
#> 98  Chr04 594886 snp12933
#> 99  Chr02 689662 snp06087
#> 100 Chr08 424912 snp26128
#> 101 Chr04 303678 snp11919
#> 102 Chr06 400385 snp19073
#> 103 Chr06 101322 snp18504
#> 104 Chr09 598852 snp30277
#> 105 Chr03 944800 snp10656
#> 106 Chr04 208264 snp11550
#> 107 Chr01 249868 snp01062
#> 108 Chr04 161865 snp11416
#> 109 Chr10 151285 snp32031
#> 110 Chr08 325690 snp25864
#> 111 Chr10 592197 snp33579
#> 112 Chr05 529444 snp16022
#> 113 Chr05  46895 snp14399
#> 114 Chr09 864811 snp31011
#> 115 Chr05 669010 snp16647
#> 116 Chr03 244905 snp07954
#> 117 Chr06 181329 snp18682
#> 118 Chr06 857999 snp20750
#> 119 Chr06 546386 snp19692
#> 120 Chr09 631892 snp30346
#> 121 Chr03 245347 snp07955
#> 122 Chr07 786798 snp23885
#> 123 Chr08 194053 snp25286
#> 124 Chr05 410847 snp15596
#> 125 Chr08 381657 snp26030
#> 126 Chr05 965848 snp17989
#> 127 Chr02 889885 snp06647
#> 128 Chr01 722910 snp03006
#> 129 Chr06 514151 snp19540
#> 130 Chr05 982603 snp18054
#> 131 Chr08 875925 snp27592
#> 132 Chr05 955966 snp17943
#> 133 Chr07 966320 snp24461
#> 134 Chr03 255953 snp07975
#> 135 Chr03 177073 snp07683
#> 136 Chr02 143882 snp04425
#> 137 Chr10 169093 snp32116
#> 138 Chr02 614741 snp05825
#> 139 Chr06 821960 snp20550
#> 140 Chr05 925365 snp17812
#> 141 Chr08 259320 snp25607
#> 142 Chr08 597985 snp26584
#> 143 Chr04 534718 snp12670
#> 144 Chr10 882103 snp34704
#> 145 Chr01 855687 snp03475
#> 146 Chr08 221528 snp25435
#> 147 Chr06 237687 snp18783
#> 148 Chr03  53335 snp07367
#> 149 Chr03 661637 snp09553
#> 150 Chr10 158248 snp32057
#> 151 Chr05 770355 snp17152
#> 152 Chr04 540383 snp12702
#> 153 Chr03 517145 snp08952
#> 154 Chr06  30923 snp18220
#> 155 Chr01 203426 snp00826
#> 156 Chr09 615417 snp30304
#> 157 Chr04 380185 snp12184
#> 158 Chr05 667756 snp16630
#> 159 Chr05 407070 snp15583
#> 160 Chr08 480046 snp26231
#> 161 Chr04 251459 snp11699
#> 162 Chr03 134021 snp07553
#> 163 Chr08 650882 snp26755
#> 164 Chr04 682794 snp13224
#> 165 Chr09 160505 snp28616
#> 166 Chr01 149086 snp00602
#>  [ reached 'max' / getOption("max.print") -- omitted 3167 rows ]

exampleData$snpEffects

#>      snp00006      snp00009      snp00011      snp00018      snp00026 
#>  0.0511092981  0.0809923308  0.1835325942  0.0281943392  0.1441369461 
#>      snp00031      snp00035      snp00036      snp00049      snp00052 
#> -0.0445462162 -0.0651931075  0.0437847860  0.1526437258 -0.1980291040 
#>      snp00075      snp00076      snp00087      snp00101      snp00102 
#>  0.0665151709 -0.2310826157  0.0084110491  0.0164703847 -0.0295777180 
#>      snp00103      snp00111      snp00113      snp00116      snp00117 
#>  0.1579919867 -0.0737983392  0.0452859755  0.0229546499  0.0414864846 
#>      snp00130      snp00131      snp00141      snp00144      snp00145 
#> -0.0878088528  0.1356897426  0.0775463603  0.0864863646  0.0726752564 
#>      snp00166      snp00172      snp00173      snp00177      snp00185 
#> -0.0685234797 -0.0249297979 -0.1061030613 -0.1209572005  0.0449026132 
#>      snp00190      snp00191      snp00201      snp00202      snp00205 
#>  0.0213965332  0.0933290496 -0.0496949112  0.1096178133  0.0663504340 
#>      snp00227      snp00239      snp00249      snp00255      snp00256 
#> -0.0094035705 -0.0740913817  0.0285733964 -0.0698192324 -0.0454730832 
#>      snp00259      snp00324      snp00349      snp00439      snp00472 
#>  0.0023972753  0.0084016640  0.1294134282  0.0960701742  0.0910269415 
#>      snp00546      snp00552      snp00562      snp00563      snp00565 
#>  0.0138610540 -0.0979172236 -0.1018755160  0.0225816003 -0.1050415213 
#>      snp00583      snp00585      snp00589      snp00594      snp00602 
#> -0.0233736572  0.1116237127 -0.1303331991 -0.0267075397 -0.0492259944 
#>      snp00603      snp00621      snp00629      snp00634      snp00640 
#>  0.1324052750  0.0440225508  0.2616430444 -0.1126755701 -0.1091948044 
#>      snp00642      snp00643      snp00655      snp00662      snp00669 
#> -0.1090436410 -0.0127361353  0.0806324269  0.0576184802 -0.0142090081 
#>      snp00670      snp00689      snp00694      snp00702      snp00711 
#> -0.1144691744  0.0552247035 -0.1197668648  0.0649633625  0.0767342863 
#>      snp00712      snp00715      snp00717      snp00721      snp00722 
#> -0.0722528972  0.1614753829 -0.0784577472 -0.0592880207  0.1036435935 
#>      snp00724      snp00742      snp00758      snp00766      snp00796 
#> -0.0286797846  0.0951160101  0.0117102186 -0.0253973467 -0.1394898095 
#>      snp00801      snp00808      snp00810      snp00812      snp00816 
#> -0.0818774476 -0.0187043772 -0.0590281940  0.0934282237 -0.1952539047 
#>      snp00826      snp00830      snp00832      snp00834      snp00836 
#>  0.0441232820  0.0539551885  0.2925421073 -0.0190238201 -0.0599926922 
#>      snp00842      snp00847      snp00855      snp00868      snp00875 
#> -0.1540706672  0.1604377699  0.0533740166  0.1183268461 -0.1232737636 
#>      snp00877      snp00888      snp00889      snp00891      snp00892 
#>  0.1226062301 -0.1237176041  0.0405057965  0.0150467129 -0.1036042555 
#>      snp00894      snp00913      snp00943      snp00948      snp00953 
#>  0.1039878622  0.0835227678 -0.0473740610  0.0138141379  0.1821587240 
#>      snp00955      snp00957      snp00958      snp00967      snp00979 
#> -0.0806874226  0.0298322158 -0.0189976715  0.0052162193  0.0729913128 
#>      snp00980      snp00982      snp00992      snp00996      snp01003 
#>  0.2996910607 -0.0384745495 -0.2308314476  0.0653618367  0.1341722772 
#>      snp01041      snp01047      snp01052      snp01062      snp01078 
#> -0.0720232962  0.4252728313  0.1752808578  0.1744121766  0.0189430868 
#>      snp01097      snp01101      snp01118      snp01120      snp01161 
#>  0.1455629422  0.1531503622 -0.0954083165 -0.0763298967  0.1635742241 
#>      snp01227      snp01229      snp01245      snp01246      snp01265 
#> -0.1103000155  0.0538996325 -0.0896011625  0.0674247580 -0.1222103923 
#>      snp01267      snp01271      snp01282      snp01283      snp01294 
#>  0.1126851536  0.0102064010  0.0476677719 -0.1261398216  0.0339544143 
#>      snp01319      snp01324      snp01325      snp01328      snp01330 
#>  0.1501393682  0.0662237606 -0.0330608230 -0.0473302753  0.1733483087 
#>      snp01353      snp01354      snp01362      snp01368      snp01373 
#> -0.1288354834  0.0506482435 -0.0987256798  0.0366639297  0.0183459679 
#>      snp01374      snp01382      snp01384      snp01392      snp01399 
#> -0.0580396859  0.0804034585  0.0299029072  0.0677209292 -0.0516660070 
#>      snp01404      snp01423      snp01424      snp01435      snp01436 
#> -0.1681078973 -0.2775602756  0.1052045399  0.0925904405 -0.0670068130 
#>      snp01480      snp01489      snp01499      snp01500      snp01504 
#> -0.1769327974 -0.1456034663 -0.1542269027 -0.1871357903 -0.1651621596 
#>      snp01512      snp01514      snp01516      snp01517      snp01531 
#> -0.2074756706 -0.0241904012  0.1389291315  0.0522867955 -0.1167742230 
#>      snp01535      snp01536      snp01540      snp01549      snp01563 
#>  0.1763187334 -0.0021643974  0.1672832712  0.0612490026 -0.1484622043 
#>      snp01599      snp01620      snp01648      snp01651      snp01657 
#> -0.1001588572  0.0035953056 -0.1246760138  0.1386533507 -0.0167374221 
#>      snp01668      snp01683      snp01686      snp01687      snp01688 
#>  0.1294826856  0.0393202834  0.2012954696 -0.0033244450 -0.0995213572 
#>      snp01697      snp01712      snp01714      snp01733      snp01779 
#> -0.2226442117 -0.0132471149 -0.0600947297  0.0171907843  0.0539784622 
#>      snp01790      snp01798      snp01808      snp01831      snp01838 
#>  0.0543420090 -0.0742097452 -0.1129740804  0.0962005029 -0.2657897086 
#>      snp01842      snp01843      snp01855      snp01859      snp01871 
#>  0.0604623569 -0.0101065297  0.1106808059 -0.1384876499 -0.0568149152 
#>      snp01877      snp01878      snp01882      snp01885      snp01894 
#> -0.0651559460 -0.0006065950  0.1046161213  0.0870378429  0.0514597657 
#>      snp01905      snp01907      snp01917      snp01930      snp01935 
#> -0.0225209010  0.0933356792 -0.0403935470 -0.0008374442  0.2082765865 
#>      snp01936      snp01937      snp01954      snp01957      snp01958 
#> -0.0167384990  0.2462467896  0.0559664425 -0.1125334416  0.0028435450 
#>      snp01987      snp01992      snp02017      snp02021      snp02024 
#>  0.0166608949 -0.0608786804  0.0722988207  0.0448778432  0.0333539684 
#>      snp02029      snp02036      snp02044      snp02049      snp02067 
#>  0.1566029460  0.2052275872  0.0719740014  0.1356659313 -0.1228592651 
#>      snp02072      snp02074      snp02121      snp02126      snp02130 
#> -0.0045853604  0.0654503311  0.0247228319 -0.0279140536  0.0767531815 
#>      snp02139      snp02140      snp02146      snp02149      snp02150 
#> -0.0109058839 -0.0404280124 -0.0262782455 -0.0808070302 -0.0271254967 
#>      snp02152      snp02159      snp02160      snp02167      snp02178 
#> -0.0134053426  0.1826608032 -0.2088200243 -0.0666830698  0.0712254898 
#>      snp02185      snp02190      snp02196      snp02206      snp02214 
#> -0.1846618819 -0.0099119862 -0.0460799858 -0.0360879178 -0.0387375916 
#>      snp02217      snp02225      snp02228      snp02231      snp02232 
#> -0.1287312475  0.0933909999  0.1815893050  0.1369104650  0.0183558697 
#>      snp02234      snp02240      snp02249      snp02261      snp02262 
#> -0.1328746584  0.0234449016 -0.1287145117 -0.0558970608  0.3216931988 
#>      snp02303      snp02328      snp02347      snp02352      snp02375 
#> -0.0609446560 -0.2097641048  0.0105003081 -0.0436889061  0.0064628441 
#>      snp02384      snp02399      snp02402      snp02406      snp02411 
#> -0.1865011939 -0.0267235185 -0.0467826457  0.0751938228 -0.0063685456 
#>      snp02426      snp02463      snp02469      snp02482      snp02503 
#>  0.1560208038  0.0174996881 -0.0758600561  0.0587545630  0.1886068268 
#>      snp02506      snp02507      snp02512      snp02520      snp02523 
#> -0.0476187832 -0.1880278788 -0.1121000281 -0.1269437676 -0.0231879139 
#>      snp02535      snp02545      snp02553      snp02560      snp02569 
#> -0.0430945559  0.0999505670 -0.1178579072  0.1174919711 -0.0356223699 
#>      snp02572      snp02586      snp02590      snp02604      snp02610 
#> -0.0904095817  0.0767808342 -0.1415174072  0.0719633865  0.0146870565 
#>      snp02618      snp02666      snp02674      snp02684      snp02693 
#>  0.0846352780 -0.0045057750 -0.0793281656  0.0640682304 -0.0302368658 
#>      snp02706      snp02712      snp02715      snp02730      snp02757 
#>  0.0582121077  0.0731825576 -0.0379371058 -0.1029427943 -0.0693413343 
#>      snp02759      snp02800      snp02831      snp02850      snp02852 
#> -0.0780731459 -0.0125196230 -0.0845357245  0.0566459843  0.0667542312 
#>      snp02853      snp02859      snp02874      snp02890      snp02894 
#> -0.0906206506  0.0296120017  0.1315987714  0.0686968527  0.0720849642 
#>      snp02905      snp02916      snp02917      snp02933      snp02938 
#> -0.1706965298 -0.0999229964 -0.0760736911  0.0992795997 -0.1028575703 
#>      snp02940      snp02945      snp02946      snp02954      snp02961 
#>  0.0402674302 -0.0345127345  0.1707301255  0.1948642637  0.0127289600 
#>      snp02962      snp02996      snp03006      snp03019      snp03021 
#>  0.0720174258 -0.0225598916 -0.1679209355 -0.0039157646 -0.1932622691 
#>      snp03023      snp03052      snp03071      snp03081      snp03094 
#> -0.0817285789 -0.1422391043  0.0920370259 -0.1140859738 -0.0584738449 
#>      snp03098      snp03103      snp03105      snp03133      snp03143 
#> -0.0354566708  0.1303284502  0.1276628333  0.0796852721  0.1053290815 
#>      snp03153      snp03165      snp03167      snp03209      snp03211 
#> -0.1244813660 -0.0007392833 -0.0527761622  0.0339580547 -0.1310212536 
#>      snp03221      snp03225      snp03247      snp03248      snp03254 
#>  0.2315892257 -0.0567616434 -0.0856325939 -0.0457044760 -0.0823142207 
#>      snp03259      snp03262      snp03268      snp03270      snp03284 
#>  0.0182768879  0.0406116234  0.1653278094  0.1076974137 -0.1617258429 
#>      snp03323      snp03380      snp03384      snp03386      snp03408 
#> -0.0105505182 -0.0814084901  0.1238536314 -0.1011374391 -0.1794054266 
#>      snp03420      snp03475      snp03476      snp03488      snp03494 
#>  0.1104452059 -0.1489127114  0.0057440458 -0.1201949109 -0.0378927966 
#>      snp03503      snp03517      snp03519      snp03533      snp03542 
#>  0.0786353071  0.1586747221 -0.0640876874  0.0502951098  0.2692639796 
#>      snp03551      snp03559      snp03568      snp03573      snp03579 
#>  0.0144970796  0.1078915735 -0.1995744289 -0.1099179307 -0.0266833902 
#>      snp03580      snp03597      snp03612      snp03622      snp03626 
#> -0.0864441572 -0.0198297796 -0.0216331798 -0.0206524903  0.1380497590 
#>      snp03642      snp03647      snp03657      snp03668      snp03669 
#>  0.2152874237  0.1147253296 -0.0092164862 -0.0419055233  0.0241348995 
#>      snp03679      snp03683      snp03684      snp03700      snp03704 
#> -0.1544151294  0.0235381976  0.0691251647 -0.0198725633 -0.0062664621 
#>      snp03711      snp03728      snp03732      snp03735      snp03747 
#>  0.0937152172  0.0397622737  0.0524856234 -0.0213943501  0.0863373229 
#>      snp03751      snp03755      snp03758      snp03760      snp03763 
#>  0.0488567808 -0.1299354492  0.0123756329  0.0146830792  0.1071882003 
#>      snp03767      snp03779      snp03795      snp03820      snp03821 
#> -0.0177129471 -0.0651849840 -0.0883285861 -0.0637085788  0.1214194779 
#>      snp03833      snp03848      snp03859      snp03881      snp03891 
#> -0.0916369023  0.1205190428  0.0043194565  0.0366025694 -0.2013488618 
#>      snp03895      snp03896      snp03904      snp03910      snp03923 
#>  0.0588715744  0.0518045920  0.0972479042  0.0157347648  0.0413221834 
#>      snp03931      snp03943      snp03952      snp03956      snp03973 
#>  0.1098953407 -0.2180802451  0.1263502759 -0.3784738957 -0.0067576843 
#>      snp03974      snp03976      snp03977      snp03982      snp03994 
#>  0.1403311664 -0.0619402151  0.0058964800 -0.1196020205  0.1644478192 
#>      snp04000      snp04004      snp04007      snp04013      snp04019 
#> -0.0101403027 -0.0651954788 -0.0567541794 -0.0559923733  0.0489271467 
#>      snp04053      snp04056      snp04066      snp04070      snp04079 
#> -0.0065573680  0.0805736920 -0.0808479354  0.1745067369 -0.0306533642 
#>      snp04118      snp04133      snp04140      snp04141      snp04142 
#>  0.2344312893  0.0120315588  0.0911511352  0.1509226010 -0.0672983028 
#>      snp04152      snp04157      snp04160      snp04164      snp04218 
#> -0.0963812860  0.0209025676  0.0074685739 -0.0435150719 -0.1032201504 
#>      snp04231      snp04254      snp04262      snp04269      snp04281 
#> -0.0319794419 -0.0402839255 -0.0535742871 -0.0161789482  0.1301125076 
#>      snp04293      snp04346      snp04358      snp04359      snp04365 
#>  0.0942472232 -0.1009237880 -0.0948750785 -0.1192555590 -0.0100037915 
#>      snp04367      snp04377      snp04394      snp04395      snp04404 
#>  0.0205770384  0.2896061875  0.0017211492 -0.0846816738  0.0265903116 
#>      snp04408      snp04424      snp04425      snp04430      snp04431 
#>  0.0135492382 -0.0939716121  0.3017455233 -0.0207069438 -0.1249094522 
#>      snp04445      snp04480      snp04486      snp04487      snp04505 
#> -0.1119089310 -0.0489013497  0.0719715679  0.1044068907 -0.0172677808 
#>      snp04537      snp04538      snp04542      snp04549      snp04570 
#> -0.1566242910  0.1178492717  0.0874471608  0.0522195359  0.1983832452 
#>      snp04593      snp04669      snp04684      snp04692      snp04699 
#> -0.0249755890 -0.0686734558  0.0112302116  0.0826328403 -0.0560500350 
#>      snp04701      snp04707      snp04716      snp04718      snp04723 
#> -0.0790110506  0.2103186583  0.1463011411 -0.1031633301 -0.0045548137 
#>      snp04762      snp04767      snp04778      snp04807      snp04816 
#> -0.0672498023 -0.0674917033  0.1706163598  0.1324377278  0.1766722028 
#>      snp04818      snp04823      snp04825      snp04900      snp04903 
#>  0.0132131833 -0.0233174197  0.0818058109 -0.0657212765  0.2031892533 
#>      snp04924      snp04929      snp04939      snp04973      snp04989 
#> -0.0337781042 -0.0357871068 -0.2088037726 -0.0351640709 -0.0411788826 
#>      snp05003      snp05004      snp05026      snp05031      snp05037 
#>  0.1351264820 -0.0543593455 -0.0265630839 -0.0627551007  0.0579834493 
#>      snp05040      snp05052      snp05060      snp05066      snp05068 
#> -0.0297758611 -0.1441260282 -0.1380377476  0.0500541404 -0.1025578164 
#>      snp05071      snp05083      snp05087      snp05115      snp05120 
#>  0.1070604035 -0.1469339663  0.0039104530 -0.1216611675 -0.0805539839 
#>  [ reached getOption("max.print") -- omitted 2833 entries ]