R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(name = NULL, class = "quantitative", qtn = NULL, qtnEff = NULL)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
#> A new species has emerged: Geneticae Exempli !
#> 
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !
myTrait$gv(example_pop)
#>                  [,1]
#> Coll0001   5.39328554
#> Coll0002   0.45752269
#> Coll0003   7.89504782
#> Coll0004  -1.90032235
#> Coll0005   6.14501291
#> Coll0006   1.36250209
#> Coll0007   3.37208527
#> Coll0008   1.78145050
#> Coll0009   4.15669689
#> Coll0010  -0.02474088
#> Coll0011   2.35313493
#> Coll0012  -3.72517802
#> Coll0013  -0.33865913
#> Coll0014   8.43454209
#> Coll0015   1.26044707
#> Coll0016  -9.70382687
#> Coll0017  -2.51648312
#> Coll0018  -5.07147540
#> Coll0019  -6.19725168
#> Coll0020   1.68223440
#> Coll0021   2.65027466
#> Coll0022  -0.04610388
#> Coll0023  -9.84879010
#> Coll0024   5.71971153
#> Coll0025   2.23180574
#> Coll0026   0.63729466
#> Coll0027   3.17541013
#> Coll0028  -2.03738587
#> Coll0029  -1.72788164
#> Coll0030   1.85094036
#> Coll0031   0.99000257
#> Coll0032   4.65204998
#> Coll0033   3.56784578
#> Coll0034  -6.72293053
#> Coll0035   7.37286465
#> Coll0036   1.73409746
#> Coll0037  -3.57792815
#> Coll0038  -2.73901292
#> Coll0039  -1.62614188
#> Coll0040   2.24306439
#> Coll0041   0.31991527
#> Coll0042   3.88775325
#> Coll0043   2.70040261
#> Coll0044  -5.65015105
#> Coll0045   4.39589417
#> Coll0046   1.03983142
#> Coll0047   5.15935526
#> Coll0048   1.76671529
#> Coll0049   1.87197729
#> Coll0050  -1.57432287
#> Coll0051  -6.10902584
#> Coll0052   2.89778943
#> Coll0053   0.72714293
#> Coll0054  -2.61298149
#> Coll0055   5.05254977
#> Coll0056  -0.52844550
#> Coll0057   4.00248565
#> Coll0058   2.48147817
#> Coll0059   2.84827864
#> Coll0060  -0.70038856
#> Coll0061   2.93010013
#> Coll0062  -4.00884629
#> Coll0063  10.50771064
#> Coll0064   5.17672982
#> Coll0065   4.03520247
#> Coll0066   5.53978613
#> Coll0067   3.18036729
#> Coll0068   6.36159795
#> Coll0069  -3.22217707
#> Coll0070  -0.42079090
#> Coll0071   2.09157909
#> Coll0072  -0.53211863
#> Coll0073 -10.03075822
#> Coll0074  -4.28935477
#> Coll0075  -5.23785248
#> Coll0076   1.43174803
#> Coll0077  10.06104982
#> Coll0078   6.13192680
#> Coll0079   3.44980415
#> Coll0080   5.39798114
#> Coll0081   2.88494834
#> Coll0082   2.76021629
#> Coll0083   1.15447637
#> Coll0084   1.25310383
#> Coll0085   5.01839641
#> Coll0086   6.17159226
#> Coll0087   2.47701932
#> Coll0088   5.97318466
#> Coll0089   0.11391058
#> Coll0090   4.29765695
#> Coll0091   4.47615647
#> Coll0092  -2.02892487
#> Coll0093   6.68542677
#> Coll0094   4.68618809
#> Coll0095   2.00750906
#> Coll0096  -0.79542362
#> Coll0097   0.77359365
#> Coll0098   8.30024811
#> Coll0099   0.58185088
#> Coll0100   0.36990573