R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(name = NULL, class = "quantitative", qtn = NULL, qtnEff = NULL)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
#> A new species has emerged: Geneticae Exempli !
#> 
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !
myTrait$gv(example_pop)
#>                [,1]
#> Coll0001  -4.788113
#> Coll0002 -16.424759
#> Coll0003 -10.761256
#> Coll0004 -18.878723
#> Coll0005 -11.717274
#> Coll0006 -17.196982
#> Coll0007 -11.277409
#> Coll0008 -20.689279
#> Coll0009  -8.146475
#> Coll0010 -11.321469
#> Coll0011 -11.451287
#> Coll0012  -8.665839
#> Coll0013 -19.262362
#> Coll0014 -16.128044
#> Coll0015 -10.016597
#> Coll0016 -21.269437
#> Coll0017 -11.624435
#> Coll0018 -21.216419
#> Coll0019  -8.234235
#> Coll0020  -9.977921
#> Coll0021 -11.394881
#> Coll0022 -16.372710
#> Coll0023 -16.911574
#> Coll0024 -10.462041
#> Coll0025 -15.516545
#> Coll0026 -16.974868
#> Coll0027 -15.793029
#> Coll0028 -16.988282
#> Coll0029 -15.470368
#> Coll0030 -19.285964
#> Coll0031 -11.640818
#> Coll0032 -12.017549
#> Coll0033 -15.485549
#> Coll0034 -11.249127
#> Coll0035 -17.984539
#> Coll0036  -5.089315
#> Coll0037 -12.012811
#> Coll0038 -14.655488
#> Coll0039 -14.816766
#> Coll0040 -13.219035
#> Coll0041 -13.477717
#> Coll0042  -1.813487
#> Coll0043 -10.078774
#> Coll0044 -18.130669
#> Coll0045 -15.008768
#> Coll0046 -12.925441
#> Coll0047 -14.337239
#> Coll0048 -13.977630
#> Coll0049 -14.573755
#> Coll0050 -13.215991
#> Coll0051 -13.686465
#> Coll0052 -11.663486
#> Coll0053 -19.047142
#> Coll0054 -17.326170
#> Coll0055 -14.742562
#> Coll0056 -17.335907
#> Coll0057 -10.096428
#> Coll0058 -19.917151
#> Coll0059  -6.155750
#> Coll0060 -17.263071
#> Coll0061 -16.885454
#> Coll0062 -14.719008
#> Coll0063 -11.420548
#> Coll0064 -23.645452
#> Coll0065  -8.047539
#> Coll0066 -13.187826
#> Coll0067 -14.715177
#> Coll0068 -19.767825
#> Coll0069  -7.434411
#> Coll0070 -10.207098
#> Coll0071  -9.213759
#> Coll0072 -11.469169
#> Coll0073 -15.458758
#> Coll0074 -14.145195
#> Coll0075 -13.899633
#> Coll0076 -22.565778
#> Coll0077 -24.050536
#> Coll0078 -15.188960
#> Coll0079 -15.339306
#> Coll0080 -10.820174
#> Coll0081 -20.301783
#> Coll0082  -8.679908
#> Coll0083 -13.238306
#> Coll0084 -10.822053
#> Coll0085 -17.074259
#> Coll0086 -10.011427
#> Coll0087 -11.456779
#> Coll0088  -6.026500
#> Coll0089 -21.185896
#> Coll0090 -19.912532
#> Coll0091 -22.938438
#> Coll0092 -12.267736
#> Coll0093  -9.977843
#> Coll0094 -18.332817
#> Coll0095 -15.285591
#> Coll0096  -9.345638
#> Coll0097 -16.431014
#> Coll0098 -10.815965
#> Coll0099  -7.629572
#> Coll0100  -6.367664