R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(name = NULL, class = "quantitative", qtn = NULL, qtnEff = NULL)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
#> A new species has emerged: Geneticae Exempli !
#> 
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !
myTrait$gv(example_pop)
#>                 [,1]
#> Coll0001  -2.4494309
#> Coll0002   1.1660152
#> Coll0003  -3.2653403
#> Coll0004  -6.4642035
#> Coll0005  -1.3107775
#> Coll0006   4.9107298
#> Coll0007  -5.0622150
#> Coll0008  -3.0406557
#> Coll0009  -5.3725108
#> Coll0010  -8.2672498
#> Coll0011  -7.7932544
#> Coll0012  -9.8576416
#> Coll0013  -6.5029224
#> Coll0014  -3.0236483
#> Coll0015  -5.4075686
#> Coll0016   0.3648288
#> Coll0017   0.9882802
#> Coll0018  -7.1310035
#> Coll0019  -6.7931235
#> Coll0020  -6.2927164
#> Coll0021 -10.4019931
#> Coll0022  -5.5657272
#> Coll0023  -7.0083236
#> Coll0024   2.2041526
#> Coll0025  -9.3694362
#> Coll0026  -1.2637294
#> Coll0027   2.3533623
#> Coll0028  -1.0154947
#> Coll0029   0.8509191
#> Coll0030   0.8947246
#> Coll0031  -5.4151128
#> Coll0032  -7.6236695
#> Coll0033   0.4281630
#> Coll0034  -2.0282837
#> Coll0035  -1.9377063
#> Coll0036 -10.2402637
#> Coll0037  -4.6053198
#> Coll0038  -3.5727651
#> Coll0039 -13.1325764
#> Coll0040   0.2468667
#> Coll0041  -3.2292620
#> Coll0042  -9.1221371
#> Coll0043  -4.5608492
#> Coll0044  -0.2474819
#> Coll0045  -6.1651324
#> Coll0046  -2.6848883
#> Coll0047  -7.6120542
#> Coll0048  -0.3694952
#> Coll0049  -5.8607029
#> Coll0050  -2.2697582
#> Coll0051  -0.3700430
#> Coll0052  -7.5703652
#> Coll0053  -5.8334659
#> Coll0054 -12.3195382
#> Coll0055  -3.4952903
#> Coll0056 -10.4163966
#> Coll0057  -7.6005761
#> Coll0058  -7.0035219
#> Coll0059  -1.7112949
#> Coll0060 -10.9683731
#> Coll0061  -3.6887829
#> Coll0062  -8.0601392
#> Coll0063  -0.1237588
#> Coll0064  -4.3694914
#> Coll0065  -1.0854893
#> Coll0066  -3.7101180
#> Coll0067  -0.2382373
#> Coll0068   2.7501445
#> Coll0069 -17.3334385
#> Coll0070   5.7925726
#> Coll0071  -0.8606216
#> Coll0072   0.7494505
#> Coll0073  -0.3336809
#> Coll0074  -0.7455309
#> Coll0075  -3.0486693
#> Coll0076  -1.8688916
#> Coll0077  -7.5444255
#> Coll0078  -9.7298368
#> Coll0079  -7.4444596
#> Coll0080   1.5957071
#> Coll0081  -7.8134391
#> Coll0082  -4.3379328
#> Coll0083  -8.6603997
#> Coll0084  -1.6653726
#> Coll0085   0.6503812
#> Coll0086  -1.8330704
#> Coll0087   2.9204795
#> Coll0088 -13.9365079
#> Coll0089 -13.3415892
#> Coll0090 -10.4955169
#> Coll0091  -2.4387032
#> Coll0092   5.3448152
#> Coll0093  -1.2978907
#> Coll0094  -3.2906051
#> Coll0095   0.9557794
#> Coll0096   2.8352785
#> Coll0097 -11.1164525
#> Coll0098  -6.2644817
#> Coll0099  -2.7807347
#> Coll0100  -1.5508588