R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(name = NULL, class = "quantitative", qtn = NULL, qtnEff = NULL)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
#> A new species has emerged: Geneticae Exempli !
#> 
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !
myTrait$gv(example_pop)
#>                  [,1]
#> Coll0001   2.41176677
#> Coll0002  -2.27905917
#> Coll0003   1.09472098
#> Coll0004  -3.68252699
#> Coll0005  -2.73603537
#> Coll0006  -2.05946617
#> Coll0007  -4.13210174
#> Coll0008  -5.50442244
#> Coll0009  -6.73461771
#> Coll0010  -0.44014719
#> Coll0011  -4.77857429
#> Coll0012   0.93047130
#> Coll0013  -5.43028480
#> Coll0014  -6.74884054
#> Coll0015  -5.78044867
#> Coll0016  -0.02019253
#> Coll0017  -5.06294008
#> Coll0018  -6.51869218
#> Coll0019 -14.33152057
#> Coll0020  -6.14229293
#> Coll0021  -1.94400013
#> Coll0022  -0.41513769
#> Coll0023   0.53712741
#> Coll0024  -3.04986175
#> Coll0025  -0.70855049
#> Coll0026  -5.87153368
#> Coll0027   3.57649210
#> Coll0028  -4.43981628
#> Coll0029 -10.97514595
#> Coll0030 -11.35041501
#> Coll0031  -9.37486896
#> Coll0032  -3.19683462
#> Coll0033  -5.64640945
#> Coll0034 -10.65432855
#> Coll0035   1.09213245
#> Coll0036  -2.62348370
#> Coll0037  -1.43985000
#> Coll0038  -7.02654168
#> Coll0039   0.50489426
#> Coll0040  -5.14707748
#> Coll0041  -2.14803011
#> Coll0042 -10.11841983
#> Coll0043  -8.74312622
#> Coll0044   0.49045202
#> Coll0045  -6.62923240
#> Coll0046   0.73956889
#> Coll0047  -4.97790945
#> Coll0048 -10.63210150
#> Coll0049   1.07356400
#> Coll0050 -10.86083886
#> Coll0051  -4.87703916
#> Coll0052  -1.18257424
#> Coll0053  -0.37261765
#> Coll0054  -7.13536588
#> Coll0055  -6.16831143
#> Coll0056  -5.21602166
#> Coll0057  -1.27370293
#> Coll0058  -7.71747137
#> Coll0059  -4.24961743
#> Coll0060   5.40869437
#> Coll0061  -4.13291508
#> Coll0062 -12.47336872
#> Coll0063  -2.14092484
#> Coll0064   1.10417549
#> Coll0065  -5.21249669
#> Coll0066  -8.51476443
#> Coll0067 -14.66962805
#> Coll0068 -12.22930185
#> Coll0069   1.42945332
#> Coll0070  -8.97619913
#> Coll0071  -9.58111894
#> Coll0072  -1.69696271
#> Coll0073  -2.51374095
#> Coll0074  -1.31305526
#> Coll0075  -1.00673984
#> Coll0076 -10.99098903
#> Coll0077  -0.08566591
#> Coll0078  -5.94074194
#> Coll0079  -2.73609055
#> Coll0080  -6.63521655
#> Coll0081  -9.21719213
#> Coll0082   1.22226675
#> Coll0083   2.52377434
#> Coll0084   0.79520779
#> Coll0085  -8.97184135
#> Coll0086  -6.44514961
#> Coll0087  -2.84525969
#> Coll0088  -6.71884490
#> Coll0089 -12.07769643
#> Coll0090   0.37953662
#> Coll0091  -3.43771777
#> Coll0092   6.69710055
#> Coll0093  -2.61856310
#> Coll0094  -2.91100597
#> Coll0095  -5.51982620
#> Coll0096  -3.72991319
#> Coll0097  -4.12858355
#> Coll0098  -2.60082481
#> Coll0099   2.11317181
#> Coll0100  -3.58238869