R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(
  name = NULL,
  class = "quantitative",
  qtn = NULL,
  qtnEff = NULL,
  vg = NULL,
  afs = NULL
)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     specName = "Geneticae Exempli",
                     ploidy = 2,
                     recombRate = 3/10^6)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     specName = "Geneticae Exempli",
                     ploidy = 2,
                     recombRate = 3/10^6)

#> A new species has emerged: Geneticae Exempli !
#> 

SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")

#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !

myTrait$gv(example_pop)

#>                 [,1]
#> Coll0001   2.7178022
#> Coll0002  -7.3329499
#> Coll0003  -4.1501852
#> Coll0004  -8.3372600
#> Coll0005  -1.8441642
#> Coll0006  -4.0918945
#> Coll0007  -3.6628397
#> Coll0008  -4.4268337
#> Coll0009  -5.5224824
#> Coll0010 -11.7017875
#> Coll0011  -8.8740015
#> Coll0012  -3.6034233
#> Coll0013   1.6110625
#> Coll0014  -3.3700490
#> Coll0015 -14.5840378
#> Coll0016  -5.3113940
#> Coll0017   1.6212251
#> Coll0018  -4.9999024
#> Coll0019   0.8756595
#> Coll0020   0.7398596
#> Coll0021   3.3453433
#> Coll0022  -4.7799026
#> Coll0023 -10.3881525
#> Coll0024   0.4525894
#> Coll0025  -5.5052386
#> Coll0026  -2.1048737
#> Coll0027   3.6202163
#> Coll0028  -5.2853341
#> Coll0029   1.9311643
#> Coll0030  -1.4158721
#> Coll0031  -1.2596829
#> Coll0032   3.6101593
#> Coll0033   1.3874153
#> Coll0034  -3.8616874
#> Coll0035  -3.2654685
#> Coll0036 -12.2465223
#> Coll0037  -6.1403798
#> Coll0038  -0.1519302
#> Coll0039  -6.3444200
#> Coll0040  -1.8021201
#> Coll0041 -17.3238240
#> Coll0042  -5.3367852
#> Coll0043  -2.5954757
#> Coll0044  -4.5675512
#> Coll0045  -9.9258339
#> Coll0046 -12.0442862
#> Coll0047   1.1812372
#> Coll0048   6.5132270
#> Coll0049  -4.7611605
#> Coll0050  -7.1648308
#> Coll0051  -7.9979857
#> Coll0052  -4.9671098
#> Coll0053  -0.4858822
#> Coll0054   0.4836165
#> Coll0055  -4.4505183
#> Coll0056  -9.6203308
#> Coll0057  -3.8084546
#> Coll0058 -11.4741095
#> Coll0059  -9.8558876
#> Coll0060  -2.9118392
#> Coll0061   2.7959231
#> Coll0062  -0.2015518
#> Coll0063  -5.3976787
#> Coll0064  -7.2025390
#> Coll0065   0.4722737
#> Coll0066   1.6360332
#> Coll0067   6.7475464
#> Coll0068   1.9196162
#> Coll0069 -10.7337546
#> Coll0070  -5.2331585
#> Coll0071  -5.4624479
#> Coll0072   5.4495686
#> Coll0073  -1.8725235
#> Coll0074  -3.1665783
#> Coll0075  -7.0840061
#> Coll0076  -1.2823682
#> Coll0077  -5.4422462
#> Coll0078  -5.8541814
#> Coll0079  -9.1301785
#> Coll0080 -11.8489319
#> Coll0081  -0.9426233
#> Coll0082  -4.7687268
#> Coll0083   2.6007378
#> Coll0084  -6.4760691
#> Coll0085  -1.2346903
#> Coll0086  -8.0501361
#> Coll0087  -2.5263611
#> Coll0088  -2.8387750
#> Coll0089  -3.1069742
#> Coll0090  -2.5833045
#> Coll0091   2.4134636
#> Coll0092  -4.3248956
#> Coll0093  -0.4692350
#> Coll0094  -3.0899308
#> Coll0095  -8.0044591
#> Coll0096  -9.4520829
#> Coll0097  -0.1481705
#> Coll0098  -5.8151304
#> Coll0099  -9.2678015
#> Coll0100  -9.8788327