R6 class representing a phenotypic trait

trait object store information about a phenotypic trait

Public fields

name

[string] Name of the trait

class

"qualitative" or "quantitative" trait ("qualitative" not implemented yet)

qtn

[character vector] names of the causal quantitative trait nucleotides (length(qtn) == length(qtnEff) must be true)

qtnEff

[numeric vector] quantitative trait nucleotides effects

Methods

Public methods

trait$new()

trait$gv()

trait$print()

trait$clone()

Method new()

Create a new trait object.

Usage

trait$new(
  name = NULL,
  class = "quantitative",
  qtn = NULL,
  qtnEff = NULL,
  vg = NULL,
  afs = NULL
)

Arguments

name

[character] name of the trait

class

"quantitative" or "qualitative"

qtn

[character vector] list of the quantitative trait nucleotides names implied in the trait

qtnEff

[numeric vector] quantitative trait nucleotides effects (see details for more information).

Details

qtn and qtnEff must be in the same order. qtnEff[n] must be the QTN effect of the QTN qtn[n].

Returns

A new `trait` object.

Examples

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli",
                     ploidy = 2)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

Method gv()

Calculate the genetic values of a population

Usage

trait$gv(pop)

Arguments

pop

[population class] population (see: population)

Examples

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)

Method print()

Display information about the object

Usage

trait$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

trait$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli",
                     ploidy = 2)

#> A new species has emerged: Geneticae Exempli !
#> 

SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")

#> Create population: Initialisation...
#> Create population: Create individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> Create population: Create population object...
#> Create population: Add individuals...
#> 1%2%3%4%5%6%7%8%9%10%11%12%13%14%15%16%17%18%19%20%21%22%23%24%25%26%27%28%29%30%31%32%33%34%35%36%37%38%39%40%41%42%43%44%45%46%47%48%49%50%51%52%53%54%55%56%57%58%59%60%61%62%63%64%65%66%67%68%69%70%71%72%73%74%75%76%77%78%79%80%81%82%83%84%85%86%87%88%89%90%91%92%93%94%95%96%97%98%99%100%
#> A new population created: Example population !

myTrait$gv(example_pop)

#>                [,1]
#> Coll0001 -4.2158834
#> Coll0002  8.3548092
#> Coll0003  3.7723037
#> Coll0004  3.8181822
#> Coll0005 12.3965430
#> Coll0006 -4.4802869
#> Coll0007  2.3991851
#> Coll0008 12.0305083
#> Coll0009  1.0671912
#> Coll0010 -2.8938691
#> Coll0011 -0.2320498
#> Coll0012  1.4804435
#> Coll0013 -3.9327907
#> Coll0014 -1.1846558
#> Coll0015 16.2560498
#> Coll0016  3.1415145
#> Coll0017 -0.1058846
#> Coll0018  3.4286479
#> Coll0019  4.3406178
#> Coll0020  6.7881144
#> Coll0021  6.5257012
#> Coll0022  7.8131965
#> Coll0023  6.3831769
#> Coll0024  4.0374779
#> Coll0025  8.3855127
#> Coll0026  5.2369238
#> Coll0027 -6.3388320
#> Coll0028 -2.9164099
#> Coll0029  2.1408798
#> Coll0030 -2.2292685
#> Coll0031  5.8138775
#> Coll0032  4.9860527
#> Coll0033 -1.9387855
#> Coll0034  1.4150423
#> Coll0035  2.0259025
#> Coll0036  4.0077140
#> Coll0037  8.0265737
#> Coll0038 -0.3183030
#> Coll0039 12.7400834
#> Coll0040  8.3389574
#> Coll0041 17.8613660
#> Coll0042  0.7842534
#> Coll0043 12.7257600
#> Coll0044 10.5740236
#> Coll0045  8.4581743
#> Coll0046  9.4841849
#> Coll0047  0.6052785
#> Coll0048 10.7164510
#> Coll0049  4.4981076
#> Coll0050  8.9119367
#> Coll0051  1.3075987
#> Coll0052 -1.8619359
#> Coll0053  4.3766028
#> Coll0054  2.7569067
#> Coll0055  2.8419351
#> Coll0056 -7.4289293
#> Coll0057 10.2420323
#> Coll0058  3.0596717
#> Coll0059  3.0606273
#> Coll0060  8.5380842
#> Coll0061 -6.1442407
#> Coll0062 -6.4541820
#> Coll0063  1.8867622
#> Coll0064  3.5846764
#> Coll0065 -3.1316436
#> Coll0066  1.4609023
#> Coll0067  2.0365347
#> Coll0068  9.1110817
#> Coll0069  3.6662765
#> Coll0070  8.5527853
#> Coll0071 -2.3053222
#> Coll0072  5.9575512
#> Coll0073  1.0461343
#> Coll0074  7.8330840
#> Coll0075  1.8686989
#> Coll0076  2.7073347
#> Coll0077  7.5065292
#> Coll0078  6.8937198
#> Coll0079  4.8387148
#> Coll0080  7.3985669
#> Coll0081  8.1122827
#> Coll0082  0.3961473
#> Coll0083 15.3687560
#> Coll0084  1.0182599
#> Coll0085  9.3036826
#> Coll0086  0.9711091
#> Coll0087 10.5160808
#> Coll0088  7.9832562
#> Coll0089  8.5707037
#> Coll0090 -2.0025086
#> Coll0091  9.4172620
#> Coll0092  3.5657786
#> Coll0093  6.3781416
#> Coll0094 -6.6802091
#> Coll0095  9.2803725
#> Coll0096  3.8545229
#> Coll0097  5.8371670
#> Coll0098  7.7040458
#> Coll0099  7.3843881
#> Coll0100  6.3786784